Match CameraModeCannon_update (drop spurious extsh)
Store the float yaw expression directly into the s16 rotX field
without the (s16)(s32) casts so the fctiwz->sth conversion lands
narrow, matching target (#83c). CameraModeCannon_update 98.89 -> 100%.

(cherry picked from commit a9a30ebb89b1ddc2d0d571a4a85687f226a233d0)

+++ src/main/dll/dll_0051_cameramodecannon.c
@@ -226,7 +226,7 @@ void CameraModeCannon_update(u8* obj)
     }
     yaw = camera->anim.rotX;
     delta = (s16)((0x8000 - lbl_803DD5A0->target->anim.rotX) - vec[1] - yaw);
-    camera->anim.rotX = (s16)(s32)((f32)(s32)yaw + (f32)(s32)delta / lbl_803E1AE0);
+    camera->anim.rotX = (f32)(s32)yaw + (f32)(s32)delta / lbl_803E1AE0;
     camera->anim.localPosX =
         lbl_803DD5A0->target->anim.localPosX -
         lbl_803E1AE4 * mathSinf(lbl_803E1AE8 * (f32)(s32)(-camera->anim.rotX) / lbl_803E1AEC);
